codechef monday

+++ codechef_251/Make_Distinct.cpp
@@ -23,7 +23,7 @@ ll solve() {
     ll n;
     cin >> n;
     vll a(n);
-    for(ll i=0; i<n; i++) cin>>a[i];
+    for(int i=0; i<n; i++) cin>>a[i];
     // code
 
     return 0;
